Updated sample
